feat(nginx-module) Add support for building lua-kong-nginx-module
The lua-kong-nginx-module nginx module adds support for mTLS amongst
other things. I also fixed the openresty version for kong enterprise
1.3. Notably, openresty 1.5.8.1 makes building libpcre a pain

https://github.com/openresty/lua-resty-core/issues/258

The workaround is to add some custom linker commands. For now I'm
not going to do that, but maybe I'll add some sort of
version-specific linker option in the future.