feat(dumpinfo) Add --dump-info options for dumping config json
In some circumstances, a user may want to have the same
configuration for some operations as their kong environment, such
as if they want to sync down a particular tag of community
edition. In that case, the new --dump-info tag dumps out the
configuration kong-env uses for setting up its environment,
allowing tooling to use the same versions and such.

Fixes #1