bugfix(module) actually install the kong nginx module lua files
The kong nginx lua module consists of two components, the actual
nginx c module that is compiled statically into openresty, and
a companion lua file which is used for interfacing with lua code.
Before now we installed the c module, but not the lua file,
leading to failures to load pretty much anything with kong 1.3.
This change rectifies that.